+++ CHANGELOG.md
@@ -10,6 +10,7 @@ dd mm yyyy
 - Docker image (prod): Added missing php.ini file (copy of php.ini-production).
 - Added login images to dist/img/ so they can be used in Markdown.
 - Fixed alliance and corporation search. [#38][i38]
+- Refactored director login for member tracking.
 - Small UI improvements.
 
 [i38]: https://github.com/tkhamez/neucore/issues/38